test_Django_autocomplete/requirements.txt
django~=2.2
django-autocomplete-light
Faker
django-filter
coverage
django-extensions
django-generic-m2m
django-gm2m
django-querysetsequence
django-taggit
django-tagging
mock
selenium
pytest
splinter
pytest-django
pytest-cov
pytest-splinter
tenacity
whitenoise
https://github.com/jpic/django-nested-admin/archive/refs/heads/master.zip#egg=django-nested-admin
django-debug-toolbar
sqlparse>=0.4.2 # not directly required, pinned by Snyk to avoid a vulnerability
djhacker
django-select2
django-redis
pandas
django-crispy-forms
django-widget-tweaks